pyproject.toml
[tool.poetry]
name = "takker_reborn"
version = "0.0.1"
description = "高い性能萝卜子Takker"
authors = ["风屿 <i@windis.cn>"]
license = "AGPL-v3"

[tool.poetry.dependencies]
python = "^3.9"
nonebot2 = "^2.0.0-alpha.14"
nonebot-adapter-cqhttp = "^2.0.0-alpha.14"
nonebot-plugin-apscheduler = "^0.1.2"
ujson = "^4.0.2"
tortoise-orm = {extras = ["accel"], version = "^0.17.6"}
playwright = "^1.13.1"
Pillow = "^8.3.1"
types-ujson = "^0.1.1"


[tool.poetry.dev-dependencies]
flake8 = "^3.9.2"
autopep8 = "^1.5.7"
nonebot-plugin-test = "^0.2.0"
mypy = "^0.910"
black = "^21.7b0"
pytest = "^6.2.4"

[nonebot.plugins]
plugins = ["nonebot_plugin_apscheduler"]
plugin_dirs = []

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
